Don't hardcode marks length to 2
There's an option here for `marks`, which are configurable list of marks
to track depth by. But rather than checking the cache length against the
length of this array, it's being assumed at a length of 2... So no more
than 2 marks will actually get tracked.

This fixes that by checking the actual length of the array instead.